add classes for generate and parse
